+++ rocAL/include/loaders/image/image_read_and_decode.h
@@ -69,6 +69,7 @@ class ImageReadAndDecode {
 
     //! returns timing info or other status information
     Timing timing();
+    size_t last_batch_padded_size();
 
    private:
     std::vector<std::shared_ptr<Decoder>> _decoder;