+++ CHANGELOG.md
@@ -16,15 +16,31 @@
 * Python support for audio decoders
 * Pytorch iterator for Audio
 * Python audio unit test, and support to verify outputs
+* rocDecode for HW decode
+* Support for Audio augmentation - PreEmphasis filter
+* Support for reading from file lists in file reader
+* Support for Audio augmentation - Spectrogram
+* Support for Audio augmentation - ToDecibels
+* Support for downmixing audio channels during decoding
+* Support for Audio augmentation - Resample
+* Support for TensorTensorAdd and TensorScalarMultiply operations
+* Support for Uniform and Normal distribution nodes
+* Support for Audio augmentation - NonSilentRegionDetection 
+* Support for generic augmentation - Slice 
+* Support for generic augmentation - Normalize
+* Support for Audio augmentation - MelFilterBank
 
 ### Optimizations
 
 * Tests
 * Setup Script
+* CentOS 7 support
+* SLES 15 SP5 support
 
 ### Changed
 
 * Image to tensor updates
+* ROCm install - use case graphics removed
 
 ### Fixed
 
@@ -36,18 +52,19 @@
   * Ubuntu - `20.04` / `22.04`
   * CentOS - `7`
   * RedHat - `8` / `9`
-  * SLES - `15-SP4`
-* ROCm: rocm-core - `5.7.0.50700-6`
+  * SLES - `15-SP5`
+* ROCm: rocm-core - `6.1.0.60100-64`
 * RPP - `rpp` & `rpp-dev`/`rpp-devel`
 * MIVisionX - `mivisionx` & `mivisionx-dev`/`mivisionx-devel`
+* rocDecode - `rocdecode` & `rocdecode-dev`/`rocdecode-devel`
 * Protobuf - `libprotobuf-dev`/`protobuf-devel`
-* RapidJSON - `rapidjson-dev` / `rapidjson-devel`
-* FFMPEG - `ffmpeg` dev package
-* OpenCV - [4.6.0](https://github.com/opencv/opencv/releases/tag/4.6.0)
-* Turbo JPEG - [Version 3.0.1](https://libjpeg-turbo.org/)
-* PyBind11 - [v2.10.4](https://github.com/pybind/pybind11)
+* RapidJSON - `https://github.com/Tencent/rapidjson`
+* Turbo JPEG - [Version 3.0.2](https://libjpeg-turbo.org/)
+* PyBind11 - [v2.11.1](https://github.com/pybind/pybind11)
+* FFMPEG - `ffmpeg 4` dev package
+* OpenCV - `libopencv` / [4.6.0](https://github.com/opencv/opencv/releases/tag/4.6.0)
 * libsndfile - [1.0.31](https://github.com/libsndfile/libsndfile/releases/tag/1.0.31)
-* rocAL Setup Script - `V2.0.1`
+* rocAL Setup Script - `V2.5.0`
 * Dependencies for all the above packages
 
 ### Known issues